rebrand to io.github.feign
