+++ RELEASE.md
@@ -1,7 +1,7 @@
-# OpenTracing Release Process
+# OpenFeign Release Process
 
 This repo uses semantic versions. Please keep this in mind when choosing version numbers.
 
-For the up-to-date release process, please refer the
-[release process from the OpenTracing Java API](https://github.com/opentracing/opentracing-java/blob/master/RELEASE.md).
+For the up-to-date release process, please refer to the
+[release process from the OpenFeign](https://github.com/OpenFeign/feign/blob/master/RELEASE.md).
