+++ feign-opentracing/src/test/java/feign/opentracing/FeignClientTracingTest.java
@@ -1,0 +1,7 @@
+package feign.opentracing;
+
+/**
+ * @author Pavol Loffay
+ */
+public class FeignClientTracingTest extends AbstractFeignTracingTest {
+}
