--- opentracing-feign/src/test/java/io/opentracing/feign/FeignClientTracingTest.java
@@ -1,7 +1,0 @@
-package io.opentracing.feign;
-
-/**
- * @author Pavol Loffay
- */
-public class FeignClientTracingTest extends AbstractFeignTracingTest {
-}
