+++ settings.xml
@@ -1,0 +1,28 @@
+<settings xmlns="http://maven.apache.org/SETTINGS/1.0.0"
+          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+          xsi:schemaLocation="http://maven.apache.org/SETTINGS/1.0.0
+                          http://maven.apache.org/xsd/settings-1.0.0.xsd">
+<servers>
+  <server>
+    <id>sonatype</id>
+    <username>${env.SONATYPE_USER}</username>
+    <password>${env.SONATYPE_PASSWORD}</password>
+  </server>
+  <server>
+    <id>bintray</id>
+    <username>${env.BINTRAY_USER}</username>
+    <password>${env.BINTRAY_KEY}</password>
+  </server>
+  <server>
+    <id>jfrog-snapshots</id>
+    <username>${env.BINTRAY_USER}</username>
+    <password>${env.BINTRAY_KEY}</password>
+  </server>
+  <server>
+    <id>github.com</id>
+    <username>${env.GH_USER}</username>
+    <password>${env.GH_TOKEN}</password>
+  </server>
+</servers>
+</settings>
+
